+++ protocol/contracts/libraries/Silo/LibLegacyWhitelist.sol
@@ -42,8 +42,8 @@ library LibLegacyWhitelist {
     function whitelistToken(
         address token,
         bytes4 selector,
-        uint16 stalkIssuedPerBdv,
-        uint24 stalkEarnedPerSeason
+        uint32 stalkIssuedPerBdv,
+        uint32 stalkEarnedPerSeason
     ) internal {
         AppStorage storage s = LibAppStorage.diamondStorage();
